Rename validate module to typing
